+++ magicgui/_qt/types.py
@@ -1,0 +1,7 @@
+"""Types for export."""
+from qtpy.QtWidgets import QPushButton
+from qtpy.QtCore import Signal
+
+
+ButtonType = QPushButton
+SignalType = Signal
