+++ magicgui/_qt/widgets/__init__.py
@@ -1,0 +1,14 @@
+"""Custom widgets for magicgui."""
+from .file_dialog import MagicFileDialog, MagicFilesDialog
+from .double_slider import QDoubleSlider
+from .data_combobox import QDataComboBox
+from .widget import WidgetType
+
+
+__all__ = [
+    "MagicFileDialog",
+    "MagicFilesDialog",
+    "QDoubleSlider",
+    "QDataComboBox",
+    "WidgetType",
+]
