+++ magicgui/_tests/test_magicgui.py
@@ -342,7 +342,8 @@ def test_unrecognized_types(qtbot):
 
     # now it should not raise an error... but `arg` should not be in the gui
     core.SKIP_UNRECOGNIZED_TYPES = True
-    gui = func.Gui()
+    with pytest.warns(UserWarning):
+        gui = func.Gui()
     assert not hasattr(gui, "arg")
     assert hasattr(gui, "b")
